Add files via upload
A side-scrolling space shooter type game made in Processing where the player can move using the arrow keys and fire using the space bar. Spawns asteroids with random sprites and enemy ships that move around and fire missiles back at the player. Not quite everything is implemented the way it should be (i.e. the enemy ships are represented by red squares), but the game is playable (in Processing) and has other neat features, like gaining health every 1,000 points the player gets, and different colored explosions for the asteroids and enemy ships.